回退：音调改回只支持整数半音（revert de265b1）
This reverts commit de265b130965eaf670c71cb5786100cd57d96a97.

滑块回 ±16 整数步进，pitch 全链路回 int（widgets/controller/params/
offline_config/offline_manager/window/change_key）。

+++ gui/infer/controller.py
@@ -13,7 +13,7 @@ logger = logging.getLogger(__name__)
 
 @dataclass
 class ModelConfig:
-    pitch: float
+    pitch: int
     gender: float
     protect: float
     f0method: str
